Enhance Sign Handling

+++ src/DB/OLEDBConnW.cpp
@@ -278,7 +278,7 @@ DB::DBConn::ConnType DB::OLEDBConn::GetConnType()
 	return DB::DBConn::CT_OLEDB;
 }
 
-Int32 DB::OLEDBConn::GetTzQhr()
+Int8 DB::OLEDBConn::GetTzQhr()
 {
 	return 0;
 }
