+++ src/Win32/WMIQueryW.cpp
@@ -92,7 +92,7 @@ DB::DBConn::ConnType Win32::WMIQuery::GetConnType()
 	return DB::DBConn::CT_WMIQUERY;
 }
 
-Int32 Win32::WMIQuery::GetTzQhr()
+Int8 Win32::WMIQuery::GetTzQhr()
 {
 	return Data::DateTime::GetLocalTzQhr();
 }
